Add TypeScript declarations for theme augmentation in MUI
- Introduced a new TypeScript declaration file for augmenting MUI's theme with additional font size and family options.
- Removed inline theme declarations from ThemeNodetool to streamline the code and maintain separation of concerns.
- Ensured existing functionality remains intact while enhancing theme customization capabilities.
